refactor: move crypto, feide and logger to lib folder
