+++ src/auth/authoptions.ts
@@ -1,7 +1,7 @@
 import 'server-only'
-import FeideProvider from './feide/FeideProvider'
+import FeideProvider from '@/lib/feide/FeideProvider'
 import VevenAdapter from './VevenAdapter'
-import { updateUserStudyProgrammes } from './feide/userRoutines'
+import { updateUserStudyProgrammes } from '@/lib/feide/userRoutines'
 import { decryptAndComparePassword } from './password'
 import prisma from '@/prisma'
 import { readPermissionsOfUser } from '@/services/permissionRoles/read'
