+++ src/auth/password.ts
@@ -1,5 +1,5 @@
 import 'server-only'
-import { hashAndEncrypter } from './crypto/hashAndEncrypter'
+import { hashAndEncrypter } from '@/lib/crypto/hashAndEncrypter'
 
 export const {
     hashAndEncrypt: hashAndEncryptPassword,
