+++ src/lib/crypto/hashAndEncrypter.ts
