+++ src/services/api-keys/update.ts
@@ -3,7 +3,7 @@ import { updateApiKeyValidation } from './validation'
 import { prismaCall } from '@/services/prismaCall'
 import { ServerError } from '@/services/error'
 import prisma from '@/prisma'
-import logger from '@/logger'
+import logger from '@/lib/logger'
 import { getOsloTime } from '@/dates/getOsloTime'
 import type { ApiKey } from '@prisma/client'
 import type { UpdateApiKeyTypes } from './validation'
