+++ src/services/education/schools/read.ts
@@ -2,7 +2,7 @@ import 'server-only'
 import { createStandardSchool } from './create'
 import { SchoolFilteredSelection, SchoolRelationIncluder } from './ConfigVars'
 import { prismaCall } from '@/services/prismaCall'
-import logger from '@/logger'
+import logger from '@/lib/logger'
 import prisma from '@/prisma'
 import { cursorPageingSelection } from '@/services/paging/cursorPageingSelection'
 import { StandardSchool } from '@prisma/client'
