+++ src/services/images/collections/read.ts
@@ -2,7 +2,7 @@ import 'server-only'
 import { specialCollectionsSpecialVisibilityMap } from './ConfigVars'
 import { readSpecialImage } from '@/services/images/read'
 import prisma from '@/prisma'
-import logger from '@/logger'
+import logger from '@/lib/logger'
 import { prismaCall } from '@/services/prismaCall'
 import { ServerError } from '@/services/error'
 import { readSpecialVisibility } from '@/services/visibility/read'
