+++ src/services/images/create.ts
@@ -5,7 +5,7 @@ import { allowedExtImageUpload, avifOptions, imageSizes } from './ConfigVars'
 import { prismaCall } from '@/services/prismaCall'
 import { createFile } from '@/services/store/createFile'
 import prisma from '@/prisma'
-import logger from '@/logger'
+import logger from '@/lib/logger'
 import sharp from 'sharp'
 import type { CreateImageTypes } from './validation'
 import type { Image, SpecialImage } from '@prisma/client'
